Add invokeAccountsSnap
