+++ packages/snaps-rpc-methods/src/permitted/invokeKeyring.ts
@@ -1,24 +1,20 @@
 import type { JsonRpcEngineEndCallback } from '@metamask/json-rpc-engine';
 import type { PermittedHandlerExport } from '@metamask/permission-controller';
 import { rpcErrors } from '@metamask/rpc-errors';
-import type {
-  InvokeKeyringParams,
-  InvokeKeyringResult,
-  InvokeSnapParams,
+import {
+  AccountsSnapHandlerType,
+  type InvokeAccountsSnapParams,
+  type InvokeKeyringParams,
+  type InvokeKeyringResult,
+  type InvokeSnapParams,
 } from '@metamask/snaps-sdk';
-import type { Snap, SnapRpcHookArgs } from '@metamask/snaps-utils';
-import { HandlerType, WALLET_SNAP_PERMISSION_KEY } from '@metamask/snaps-utils';
 import type { PendingJsonRpcResponse, JsonRpcRequest } from '@metamask/utils';
-import { hasProperty, type Json } from '@metamask/utils';
+import { isObject, type Json } from '@metamask/utils';
 
 import type { MethodHooksObject } from '../utils';
-import { getValidatedParams } from './invokeSnapSugar';
 
 const hookNames: MethodHooksObject<InvokeKeyringHooks> = {
-  hasPermission: true,
-  handleSnapRpcRequest: true,
-  getSnap: true,
-  getAllowedKeyringMethods: true,
+  invokeAccountSnap: true,
 };
 
 /**
@@ -35,17 +31,7 @@ export const invokeKeyringHandler: PermittedHandlerExport<
 };
 
 export type InvokeKeyringHooks = {
-  hasPermission: (permissionName: string) => boolean;
-
-  handleSnapRpcRequest: ({
-    snapId,
-    handler,
-    request,
-  }: Omit<SnapRpcHookArgs, 'origin'> & { snapId: string }) => Promise<unknown>;
-
-  getSnap: (snapId: string) => Snap | undefined;
-
-  getAllowedKeyringMethods: () => string[];
+  invokeAccountSnap: (params: InvokeAccountsSnapParams) => Promise<unknown>;
 };
 
 /**
@@ -58,11 +44,8 @@ export type InvokeKeyringHooks = {
  * function.
  * @param end - The `json-rpc-engine` "end" callback.
  * @param hooks - The RPC method hooks.
- * @param hooks.handleSnapRpcRequest - Invokes a snap with a given RPC request.
- * @param hooks.hasPermission - Checks whether a given origin has a given permission.
- * @param hooks.getSnap - Gets information about a given snap.
- * @param hooks.getAllowedKeyringMethods - Get the list of allowed Keyring
- * methods for a given origin.
+ * @param hooks.invokeAccountSnap - A function to invoke an account Snap designated by its parameters,
+ * bound to the requesting origin.
  * @returns Nothing.
  */
 async function invokeKeyringImplementation(
@@ -70,62 +53,21 @@ async function invokeKeyringImplementation(
   res: PendingJsonRpcResponse<InvokeKeyringResult>,
   _next: unknown,
   end: JsonRpcEngineEndCallback,
-  {
-    handleSnapRpcRequest,
-    hasPermission,
-    getSnap,
-    getAllowedKeyringMethods,
-  }: InvokeKeyringHooks,
+  { invokeAccountSnap }: InvokeKeyringHooks,
 ): Promise<void> {
-  let params: InvokeSnapParams;
   try {
-    params = getValidatedParams(req.params);
-  } catch (error) {
-    return end(error);
-  }
+    const { params } = req;
 
-  // We expect the MM middleware stack to always add the origin to requests
-  const { origin } = req as JsonRpcRequest & { origin: string };
-  const { snapId, request } = params;
+    if (!isObject(params)) {
+      throw rpcErrors.invalidParams({
+        message: 'Expected params to be a single object.',
+      });
+    }
 
-  if (!origin || !hasPermission(WALLET_SNAP_PERMISSION_KEY)) {
-    return end(
-      rpcErrors.invalidRequest({
-        message: `The snap "${snapId}" is not connected to "${origin}". Please connect before invoking the snap.`,
-      }),
-    );
-  }
-
-  if (!getSnap(snapId)) {
-    return end(
-      rpcErrors.invalidRequest({
-        message: `The snap "${snapId}" is not installed. Please install it first, before invoking the snap.`,
-      }),
-    );
-  }
-
-  if (!hasProperty(request, 'method') || typeof request.method !== 'string') {
-    return end(
-      rpcErrors.invalidRequest({
-        message: 'The request must have a method.',
-      }),
-    );
-  }
-
-  const allowedMethods = getAllowedKeyringMethods();
-  if (!allowedMethods.includes(request.method)) {
-    return end(
-      rpcErrors.invalidRequest({
-        message: `The origin "${origin}" is not allowed to invoke the method "${request.method}".`,
-      }),
-    );
-  }
-
-  try {
-    res.result = (await handleSnapRpcRequest({
-      snapId,
-      request,
-      handler: HandlerType.OnKeyringRequest,
+    res.result = (await invokeAccountSnap({
+      snapId: params.snapId,
+      request: params.request,
+      type: AccountsSnapHandlerType.Keyring,
     })) as Json;
   } catch (error) {
     return end(error);
